requirements.txt
# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
aiofiles>=23.2.1

# 认证与安全
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# 电子书解析
ebooklib>=0.18
mobi>=0.3.3

# 压缩文件处理
py7zr>=0.20.0
rarfile>=4.1
pycdlib>=1.14.0

# 工具
pyyaml>=6.0.1
loguru>=0.7.2
chardet>=5.2.0
apscheduler>=3.10.0
pillow>=10.0.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
